requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
rembg==2.0.67
Pillow==11.0.0
python-multipart==0.0.20
